+++ config.py
@@ -1,4 +1,4 @@
 GITHUB_ROOT_URL = r"https://raw.githubusercontent.com/7dJx1qP/stash-userscripts/%%BRANCH%%/"
-BUNDLE_VERSION = "0.21.1"
+BUNDLE_VERSION = "0.22.0"
 SERVER_URL = "http://localhost:9999"
 NAMESPACE = "https://github.com/7dJx1qP/stash-userscripts"