+++ src/body/Stash Set Stashbox Favorite Performers.user.js
@@ -12,7 +12,7 @@
         updateTextInput,
     } = unsafeWindow.stash;
 
-    const MIN_REQUIRED_PLUGIN_VERSION = '0.5.0';
+    const MIN_REQUIRED_PLUGIN_VERSION = '0.6.0';
 
     const TASK_NAME = 'Set Stashbox Favorite Performers';
     stash.visiblePluginTasks.push(TASK_NAME);
@@ -26,11 +26,13 @@
             alert('No Stashbox configured.');
         }
         for (const { endpoint } of data.data.configuration.general.stashBoxes) {
+            if (endpoint !== 'https://stashdb.org/graphql') continue;
             await stash.runPluginTask("userscript_functions", "Set Stashbox Favorite Performers", [{"key":"endpoint", "value":{"str": endpoint}}]);
         }
     }
 
     async function runSetStashBoxFavoritePerformerTask(endpoint, stashId, favorite) {
+        if (endpoint !== 'https://stashdb.org/graphql') return;
         return stash.runPluginTask("userscript_functions", "Set Stashbox Favorite Performer", [{"key":"endpoint", "value":{"str": endpoint}}, {"key":"stash_id", "value":{"str": stashId}}, {"key":"favorite", "value":{"b": favorite}}]);
     }