+++ src/body/Stash Tag Image Cropper.user.js
@@ -21,11 +21,11 @@
     let cropper = null;
 
     stash.addEventListener('page:tag:scenes', function () {
-        waitForElementId('tag-tabs-tab-scenes', function () {
+        waitForElementClass('detail-container', function () {
             const cropBtnContainerId = "crop-btn-container";
             if (!document.getElementById(cropBtnContainerId)) {
                 const tagId = window.location.pathname.replace('/tags/', '').split('/')[0];
-                const image = getElementByXpath("//div[contains(@class, 'logo-container')]//img[@class='logo']");
+                const image = getElementByXpath("//div[contains(@class, 'detail-header-image')]//img[@class='logo']");
                 image.parentElement.addEventListener('click', (evt) => {
                     if (cropping) {
                         evt.preventDefault();
@@ -35,11 +35,11 @@
                 const cropBtnContainer = document.createElement('div');
                 cropBtnContainer.setAttribute("id", cropBtnContainerId);
                 cropBtnContainer.classList.add('mb-2', 'text-center');
-                insertAfter(cropBtnContainer, image.parentElement);
+                image.parentElement.appendChild(cropBtnContainer);
 
                 const cropInfo = document.createElement('p');
 
-                const imageUrl = getElementByXpath("//div[contains(@class, 'logo-container')]//img[@class='logo']/@src").nodeValue;
+                const imageUrl = getElementByXpath("//div[contains(@class, 'detail-header-image')]//img[@class='logo']/@src").nodeValue;
                 const cropStart = document.createElement('button');
                 cropStart.setAttribute("id", "crop-start");
                 cropStart.classList.add('btn', 'btn-primary');