+++ src/header/Stash Batch Result Toggle.user.js
@@ -2,7 +2,7 @@
 // @name        Stash Batch Result Toggle
 // @namespace   %NAMESPACE%
 // @description Batch toggle scene tagger search result fields
-// @version     0.4.1
+// @version     0.5.0
 // @author      7dJx1qP
 // @match       %MATCHURL%
 // @grant       unsafeWindow