+++ src/header/Stash Markdown.user.js
@@ -2,7 +2,7 @@
 // @name        Stash Markdown
 // @namespace   %NAMESPACE%
 // @description Adds markdown parsing to tag description fields
-// @version     0.1.1
+// @version     0.2.0
 // @author      7dJx1qP
 // @match       %MATCHURL%
 // @grant       unsafeWindow