+++ src/header/Stash Set Stashbox Favorite Performers.user.js
@@ -2,7 +2,7 @@
 // @name        Stash Set Stashbox Favorite Performers
 // @namespace   %NAMESPACE%
 // @description Set Stashbox favorite performers according to stash favorites. Requires userscript_functions stash plugin
-// @version     0.2.1
+// @version     0.3.0
 // @author      7dJx1qP
 // @match       %MATCHURL%
 // @grant       unsafeWindow