+++ src/header/Stash StashID Icon.user.js
@@ -2,7 +2,7 @@
 // @name        Stash StashID Icon
 // @namespace   %NAMESPACE%
 // @description Adds checkmark icon to performer and studio cards that have a stashid
-// @version     0.1.2
+// @version     0.2.0
 // @author      7dJx1qP
 // @match       %MATCHURL%
 // @grant       unsafeWindow