+++ src/header/Stash Stats.user.js
@@ -2,7 +2,7 @@
 // @name        Stash Stats
 // @namespace   %NAMESPACE%
 // @description Add stats to stats page
-// @version     0.2.0
+// @version     0.3.0
 // @author      7dJx1qP
 // @match       %MATCHURL%
 // @grant       unsafeWindow