+++ src/header/Stash Tag Image Cropper.user.js
@@ -2,7 +2,7 @@
 // @name        Stash Tag Image Cropper
 // @namespace   %NAMESPACE%
 // @description Adds an image cropper to tag page
-// @version     0.1.1
+// @version     0.2.0
 // @author      7dJx1qP
 // @match       %MATCHURL%
 // @resource    IMPORTED_CSS https://raw.githubusercontent.com/fengyuanchen/cropperjs/main/dist/cropper.min.css